+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Changed
 - Minimum constraint of password change rule is never null.
+- Password change rule cannot be constructed with null minimum constraint.
 
 ### Fixed
 - Nothing